Batch-Card-Detection-App-main/requirements.txt
tensorflow==2.12.0
opencv-python==4.7.0.72
numpy==1.23.5
pillow==9.2.0
scipy==1.10.1
h5py==3.7.0
scikit-learn==1.2.2
matplotlib==3.7.1
streamlit==1.25.0
numpy>=1.21.0
tensorflow>=2.0.0
opencv-python>=4.5.0

